+++ plugins/eslint-plugin-react-x/src/rules/immutability/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed a false positive where Next.js `useRouter().push()` calls inside frozen callbacks were treated as in-place mutations of the captured router value; variable-declarator aliases of the router are recognized too.
+
 ## [5.14.3] - 2026-07-11
 
 ### Changed